- Upsert spending

+++ app/src/main/java/com/snister/carnagealpha/features/expense_tracker/data/models/SpendingDataModel.kt
@@ -12,7 +12,7 @@ data class SpendingDataModel(
 
     val spendingName: String,
 
-    val spendingAmount: Double,
+    val spendingAmount: Long,
 
     val dateTime: String,
 )
